TicketDeletePlugin:
Don't die when removing attachment changs.

Works on 2.3 again.

git-svn-id: https://trac-hacks.org/svn@585 7322e99d-02ea-0310-aa39-e9a107903beb

+++ ticketdeleteplugin/0.9/ticketdelete/web_ui.py
@@ -50,7 +50,7 @@ class TicketDeletePlugin(Component):
                             deletions = None
                             if "multidelete" in req.args:
                                 deletions = [x.split('_') for x in req.args.getlist('delete')]
-                                deletions.sort(key=lambda x: x[1], reverse=True)
+                                deletions.sort(lambda a,b: cmp(b[1],a
                             else:
                                 buttons = [x[6:] for x in req.args.keys() if x.startswith('delete')]
                                 deletions = [buttons[0].split('_')]
@@ -148,14 +148,17 @@ class TicketDeletePlugin(Component):
         cursor = db.cursor()
         ticket = Ticket(self.env,id)
         if field:
-            custom_fields = [f['name'] for f in ticket.fields if f.get('custom')]
-            if field != "comment" and not [1 for time, author, field2, oldval, newval in ticket.get_changelog() if time > ts and field == field2]:
-                oldval = [old for _, _, field2, old, _ in ticket.get_changelog(ts) if field2 == field][0]
-                if field in custom_fields:
-                    cursor.execute("UPDATE ticket_custom SET value=%s WHERE ticket=%s AND name=%s", (oldval, id, field))
-                else:
-                    cursor.execute("UPDATE ticket SET %s=%%s WHERE id=%%s" % field, (oldval, id))
-            cursor.execute("DELETE FROM ticket_change WHERE ticket = %s AND time = %s AND field = %s", (id, ts, field))
+            if field == 'attachment':
+                pass # Better handling still pending
+            else:
+                custom_fields = [f['name'] for f in ticket.fields if f.get('custom')]
+                if field != "comment" and not [1 for time, author, field2, oldval, newval in ticket.get_changelog() if time > ts and field == field2]:
+                    oldval = [old for _, _, field2, old, _ in ticket.get_changelog(ts) if field2 == field][0]
+                    if field in custom_fields:
+                        cursor.execute("UPDATE ticket_custom SET value=%s WHERE ticket=%s AND name=%s", (oldval, id, field))
+                    else:
+                        cursor.execute("UPDATE ticket SET %s=%%s WHERE id=%%s" % field, (oldval, id))
+                cursor.execute("DELETE FROM ticket_change WHERE ticket = %s AND time = %s AND field = %s", (id, ts, field))
         else:
             for _, _, field, _, _ in ticket.get_changelog(ts):
                 self._delete_change(id, ts, field)
